Extracted dataclass to fix circular dependency
