+++ src/demolition_spider.py
@@ -1,32 +1,8 @@
-from dataclasses import dataclass
 import json
-from typing import Any
 from scrapy import Spider
 from scrapy.http import FormRequest, Response
 from scrapy.utils.response import open_in_browser
-from enum import Enum
-from .sendmail import mail_sender
-
-
-class PermitType(str, Enum):
-    commercial_wrecking_permit = "Building/Wrecking/Commercial/NA"
-    residential_wrecking_permit = "Building/Wrecking/Residential/NA"
-
-
-@dataclass
-class BuildingServicesSearchResult:
-    # TODO validation (regex?)
-    record_number: str  # e.g. WRK2024R-00138
-    record_details_link: str  # URL to the record details page
-    record_type: PermitType
-    project_name: str | None
-    address: str
-    expiration_date: Any | None  # I haven't found any records w/ this yet
-    short_notes: str | None
-
-    def __str__(self):
-        return f"{self.record_type.value}:\n{self.record_number} - {self.project_name} - {self.address}"
-
+from .schema import BuildingServicesSearchResult, PermitType
 
 class DemolitionSpider(Spider):
     name = "DemolitionSpider"
@@ -135,5 +111,4 @@ class DemolitionSpider(Spider):
                     short_notes=row.xpath("string(.//td[7])").get().strip(),
                 )
             )
-        mail_sender.send_email(records)
         return records
